// src/app/modules/configuration/general/models/school-facility-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Gets or sets SchoolFacilityView
 */
export interface SchoolFacilityView { 
    /**
     * Gets or Sets Id
     */
    id: string;
    /**
     * Gets or sets Name
     */
    facilityName: string;
    /**
     * Gets or sets Code
     */
    facilityCode: string;
    /**
     * Gets or sets TypeId
     */
    typeId: number;
}